DE-4816 Add manifest refresh config and update audio
This adds a new configuration option to specify how often (at most)
manifests can be updated. This was hardcoded to 1s before but that is
not sufficient for low latency playback since we need to update
potentialy faster.

Also adds the same update logic that is used for regular media playlist
to the audio playlists (for alternative audio) where instead of the
average target it was using the acutal target, which was way too long in
the low latencey case where we need to update playlists fast and often.